Add a tool test class
